pyproject.toml
[project]
name = "zotpilot"
version = "0.1.0"
description = "AI-powered Zotero research assistant — MCP server + agent skill"
requires-python = ">=3.10"
license = "MIT"
authors = [{ name = "Xiaodong Zhuang" }]
keywords = ["zotero", "mcp", "semantic-search", "research", "ai"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]
dependencies = [
    "chromadb>=0.4.0",
    "google-genai>=1.0.0",
    "pymupdf>=1.25.0",
    "pymupdf4llm>=0.0.17",
    "pymupdf-layout",
    "fastmcp>=2.0.0",
    "pydantic>=2.0.0",
    "httpx",
    "rapidfuzz",
    "tqdm",
]

[project.urls]
Homepage = "https://github.com/xdzhuang/ZotPilot"
Repository = "https://github.com/xdzhuang/ZotPilot"
Issues = "https://github.com/xdzhuang/ZotPilot/issues"

[project.optional-dependencies]
paddle = ["paddlepaddle>=3.0.0", "paddleocr>=3.0.0"]
vision = ["anthropic>=0.40.0", "openai>=1.0.0"]
write = ["pyzotero>=1.10.0"]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0", "ruff", "mypy"]
all = ["zotpilot[paddle,vision,write]"]

[project.scripts]
zotpilot = "zotpilot.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/zotpilot"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "--cov=zotpilot --cov-report=term-missing"
